+++ src/heroes/routes/HeroesRoutes.jsx
@@ -1,6 +1,6 @@
 import { Navigate, Route, Routes } from "react-router-dom"
 import { Navbar } from "../../ui"
-import { DcPage, HeroesPage, MarvelPage, SearchPage } from "../pages"
+import { DcPage, HeroesPage, ProfilePage, SearchPage } from "../pages"
 
 export const HeroesRoutes = () => {
   return (
@@ -8,7 +8,7 @@ export const HeroesRoutes = () => {
       <Navbar />
       <div className="container">
         <Routes>
-          <Route path="perfil" element={<MarvelPage />} />
+          <Route path="perfil" element={<ProfilePage />} />
           <Route path="dc" element={<DcPage />} />
 
           <Route path="search" element={<SearchPage />} />
